+++ NEWS.md
@@ -1,6 +1,8 @@
 # tsibble dev
 
 * `index_by()` allows for grouping index variable; and `group_by()` will throw out an error for grouping index.
+* Defunct `fill_na()` in favour of `fill_gaps()`.
+* Deprecate `as.tsibble()` in favour of `as_tsibble()`.
 
 # tsibble 0.8.0